script for overlapping segments with different volumes
You should create folders called -folder1_overlap_diffvol- for accomodate the overlapped files. The first label in the combined name is the one with more volume. The folders should be created in a directory called - data/UrbanSound8K/audio_overlap_diff_volumes -

There is a parameter at the beginning of the script called reduction_vol which controls on how many db you want the second sound to be lower